requirements.txt
fastapi
uvicorn
selenium
python-jose[cryptography]
pydantic
SQLAlchemy
alembic
pytest
httpx
python-multipart